feat: Extracted get ip localization to own service
